+++ ReadingRainbow/screens/sign_up.js
@@ -51,16 +51,19 @@ export default function SignUp({navigation}) {
                     data.password
                 );
                 CreateAccount(data.username, data.email, digest).then((account_created) => {
-                    if (account_created.toLowerCase == 'true') {
-                        console.log('account_activated', account_created)
-                        ReSendEmail(data.username, data.password);
+                    if (account_created) {
+                        Alert.alert(
+                            "Account Creation Successful",
+                            "Please confirm your email by clicking the link sent to the email address you provided."
+                        );
                     } else {
                         Alert.alert(
                             "Account Creation Failed",
                             "Please try again"
                         ); 
                     }
-                }).then(navigation.navigate('SignInScreen'));
+                    navigation.navigate('SignInScreen');
+                });
             } catch (e) {
                 console.log(e);
             }
