--- ReadingRainbowAPI/ReadingRainbowAPI.csproj
@@ -1,22 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="AutoMapper" Version="10.1.1" />
-    <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="8.1.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="3.1.0" />
-    <PackageReference Include="Microsoft.IdentityModel.Tokens" Version="6.8.0" />
-    <PackageReference Include="Neo4j.Driver" Version="4.2.0" />
-    <PackageReference Include="Neo4jClient" Version="4.0.3" />
-    <PackageReference Include="RSG.Promise" Version="3.0.1" />
-    <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="6.8.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-	<None Include=".\appsettings.json" CopyToPublishDirectory="Always" CopyToOutputDirectory="Always" />
-  </ItemGroup> 
-
-</Project>
