--- ReadingRainbowAPI/appsettings.json
@@ -1,23 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
-  },
-  "AllowedHosts": "*",
-  "Access-Control-Allow-Origin": "*",
-  "JwtConfig": 
-  {  
-    "secret": "PDv7DrqznYL6nv7DrqzjnQYO9JxIsWdcjnQYL6nu0f",  
-    "expirationInMinutes": 1440
-  } , 
-  "NeO4jConnectionSettings":
-  {
-    "NeoURI": "http://localhost:7474",
-    "NeoUserName":"neo4j",
-    "NeoPW": "abc123"     
-  }
-
-}
